Added Appstream metainfo announcing HW support.
Added Appstream metainfo XML announcing the hardware
handled by this package.

Including this information in the package will ensure programs mapping
hardware to packages using Appstream information, like the isenkram
package, will know that this package is useful on machines where the USB
IDs are discovered.

The list of modaliases for supported hardware is based on the idea that
every kernel module in the drivers/edact/ subdirectory is handled by
this tool, and was generated using this oneliner with the output slightly
modified:

  for f in $(ls /lib/modules/6.9.7-amd64/kernel/drivers/edac/|sed s/.ko.xz//); do \
    echo "<modalias>lkmodule:$f</modalias>"; \
    for a in $(sudo modinfo $f|grep alias|awk '{print $2}'); do \
      echo "<modalias>$a</modalias>"; \
    done; \
  done

Please adjust the hardware mapping if this assumtion is wrong.